+++ lib/rubocop/ast/node/regexp_node.rb
@@ -32,6 +32,26 @@ module RuboCop
         children.select(&:str_type?).map(&:str_content).join
       end
 
+      # @return [Bool] if the regexp is a /.../ literal
+      def slash_literal?
+        loc.begin.source == '/'
+      end
+
+      # @return [Bool] if the regexp is a %r{...} literal (using any delimiters)
+      def percent_r_literal?
+        !slash_literal?
+      end
+
+      # @return [String] the regexp delimiters (without %r)
+      def delimiters
+        [loc.begin.source[-1], loc.end.source[0]]
+      end
+
+      # @return [Bool] if char is one of the delimiters
+      def delimiter?(char)
+        delimiters.include?(char)
+      end
+
       # @return [Bool] if regexp contains interpolation
       def interpolation?
         children.any?(&:begin_type?)